backend/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
cloudinary
python-dotenv
dnspython
bcrypt
PyJWT
gradio_client
